Validate partition and replication factor configuration values
Also changed to `short` for replication factor configurations, and moved some magic string values into constants.

+++ connect/runtime/src/main/java/org/apache/kafka/connect/storage/KafkaConfigBackingStore.java
@@ -423,7 +423,7 @@ public class KafkaConfigBackingStore implements ConfigBackingStore {
         final NewTopic topicDescription = TopicAdmin.defineTopic(topic).
                 compacted().
                 partitions(1).
-                replicationFactor(config.getInt(DistributedConfig.CONFIG_STORAGE_REPLICATION_FACTOR_CONFIG)).
+                replicationFactor(config.getShort(DistributedConfig.CONFIG_STORAGE_REPLICATION_FACTOR_CONFIG)).
                 build();
         TopicSupplier topicSupplier = new TopicSupplier() {
             @Override
